statistics collector and statistics cache

+++ src/test/java/com/example/n26/n26transactions/core/statistics/CountCalculatorTest.java
@@ -13,19 +13,19 @@ public class CountCalculatorTest {
 
   @Test
   public void shouldReturn0ForEmptyList() {
-    Integer actual = calculator.calculate(Collections.emptyList());
+    Long actual = calculator.calculate(Collections.emptyList());
     assertEquals(0.0, actual, 0.1);
   }
 
   @Test
   public void shouldReturn0ForNull() {
-    Integer actual = calculator.calculate(null);
+    Long actual = calculator.calculate(null);
     assertEquals(0.0, actual, 0.1);
   }
 
   @Test
   public void shouldReturn3For1_2_3() {
-    Integer actual = calculator.calculate(Arrays.asList(1.0, 2.0, 3.0));
-    assertEquals(Integer.valueOf(3), actual);
+    Long actual = calculator.calculate(Arrays.asList(1.0, 2.0, 3.0));
+    assertEquals(Long.valueOf(3L), actual);
   }
 }
